+++ app.js
@@ -10,7 +10,9 @@ const indexRouter = require('./routes/index');
 const usersRouter = require('./routes/users');
 const mailerRouter = require('./routes/mailerRouter');
 const authRouter = require('./routes/auth')
-const newsRouter = require('./routes/news')
+const newsRouter = require('./routes/news');
+const testimonialsRouter = require('./routes/testimonialsRouter');
+const { application } = require('express');
 
 const app = express();
 app.use(cors())
@@ -29,6 +31,7 @@ app.use('/', indexRouter);
 app.use('/users', usersRouter);
 app.use('/auth', authRouter)
 app.use('/news', newsRouter)
+app.use('/testimonials', testimonialsRouter);
 
 app.use('/api/mail', mailerRouter)
