+++ controllers/mainControllers.js
@@ -1,8 +1,5 @@
 let mockDataJson = require('./MOCK_DATA.json')
 let mockData = JSON.parse(JSON.stringify(mockDataJson))
-const db = require('../models');
-const { validationResult } = require("express-validator");
-
 
 const mainControllers = {
     index: (req, res) => {
@@ -16,23 +13,6 @@ const mainControllers = {
             return res.status(400).json({ errors: [{ msg: "No organization found" }] })
         }
     },
-    testimonials: (req, res) => {
-        const errors = validationResult(req)
-        if (!errors.isEmpty()) {
-            return res.json({errors: errors.mapped()});
-        } else {
-            db.Testimonials.create({
-                name: req.body.name,
-                content: req.body.content
-            })
-                .then((testimony) => {
-                    return res.status(200).send(testimony)
-                })
-                .catch(error => {
-                    console.error(error)
-                })
-        }
-    }
 }
 
 module.exports = mainControllers
